somewhat working (but crazy messy) ggml q4_0 loading and execution
